Add cores and total_memory parameters
These overlap with (and might possibly replace) the threads and memory
parameters.  I believe that they are more consistent with typcial use.

Now we ask people to specify the cores and total memory of a node.  This
should remain relatively static over time.  Then we allow them to tune
processes as they like independently of other parameters.