This commit adds support for a configuration parameter
This commit adds support for a configuration parameter to each of the
clsuter types in jobqueue. It does this by adding a configuration
parameter that is used to override the scheduler_name class variable.
I also demoted scheduler_name to an instance variable as you could now
theoretically have multiple types of PBS clusters.