Add a `job_extra` flag to the SGECluster
The SGECluster has no `job_extra` setting like all the other clusters do, which limits its abilities. This PR adds the `job_extra` flag in the same way as it appears in the other cluster files.

xref #219 and #235 which seemed to allude to `SGECluster` having the `job_extra` flag, but no such flag appears to exist.