New attempt at #411 (adding extra argument to condor_submit) (#514)
* Add submit_command_extra and cancel_command_extra arguments

`submit_command_extra` and `cancel_command_extra` (config options
`jobqueue.X.submit-command-extra` and `jobqueue.X.cancel-command-extra`)
are lists of strings that are arguments passed on as-is to `condor_submit`
and `condor_rm` when submitting or removing HTCondor jobs.

* Inline shell quoting for setting submit_command/cancel_command

* Missed a space

* Turn core.Job._close_job() into an instance method so we can override cancel_command in the instance

* Add hostname to htcondor test container

Some htcondor parameters (e.g. "-name" to specify a schedd name) expect
to be given an FQDN (or something that looks like one) and will fail if
not given a parameter with a `.` in it.

* Add stub test for submit_command_extra and cancel_command_extra
i# On branch pr/411-extra-commands

* Add tests for:

- htcondor with a broken submit command
- htcondor with working submit/cancel commands with extra arguments
- htcondor with a broken cancel command

* Better test for whether a remove happened

(assuming no other workers are running -- it would be even better if
I had the job IDs of the workers from the cluster.

Also don't wait so long.  A remove is quicker than a submit.

* Got the assert backwards.

* Turn core.Job._close_job() into an instance method so we can override cancel_command in the instance

* revert changing internal close_job to instance method

* add cancel_command also to finalizer and local overide

* revert to some more relaxed testing

* try different testing

* add timeout to testing

* remove broken_submit test from htcondor

* revert unrelated change

Co-authored-by: Matyas Selmeci <matyas@cs.wisc.edu>